requirements.txt
black==23.9.1
Django==4.2.5
djangorestframework==3.14.0
drf-spectacular==0.26.5
Pillow==10.0.1
pip==22.0.2
psycopg2==2.9.9
pytest==6.2.5
python-dotenv==1.0.0
supabase==1.2.0
